Removed unneeded classes. Moved all dry update functionality into preexisting classes.
