+++ GitDepend.UnitTests/GitDepend.UnitTests.csproj
@@ -119,7 +119,7 @@
     <Compile Include="Visitors\PullBranchVisitorTests.cs" />
     <Compile Include="Visitors\PushBranchVisitorTests.cs" />
     <Compile Include="Visitors\RemoveDependencyVisitorTests.cs" />
-    <Compile Include="Visitors\VerifyCorrectBranchDryVisitorTests.cs" />
+    <Compile Include="Visitors\VerifyCorrectBranchVisitorTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
