+++ GitDepend/Commands/UpdateCommand.cs
@@ -43,15 +43,7 @@ namespace GitDepend.Commands
         /// <returns>The return code.</returns>
         public ReturnCode Execute()
         {
-			NamedDependenciesVisitor verifyVisitor;
-	        if (!_options.Dry)
-	        {
-				verifyVisitor = new VerifyCorrectBranchVisitor(_options.Dependencies);
-			}
-	        else
-	        {
-				verifyVisitor =  new VerifyCorrectBranchDryVisitor(_options.Dependencies);
-	        }
+            var verifyVisitor = new VerifyCorrectBranchVisitor(_options.Dependencies, _options.Dry);
             _algorithm.TraverseDependencies(verifyVisitor, _options.Directory);
 
             if (verifyVisitor.ReturnCode != ReturnCode.Success)
@@ -71,11 +63,11 @@ namespace GitDepend.Commands
 
             if (checkArtifactsVisitor.ReturnCode == ReturnCode.Success && checkArtifactsVisitor.UpToDate)
             {
-	            if (!_options.Dry)
-	            {
-		            _console.WriteLine(strings.PACKAGES_UP_TO_DATE);
-		            return ReturnCode.Success;
-	            }
+                if (!_options.Dry)
+                {
+                    _console.WriteLine(strings.PACKAGES_UP_TO_DATE);
+                    return ReturnCode.Success;
+                }
             }
 
             if (checkArtifactsVisitor.ReturnCode != ReturnCode.Success)
@@ -88,80 +80,83 @@ namespace GitDepend.Commands
 
             _console.WriteLine();
 
-	        if (!_options.Dry)
-	        {
-		        _algorithm.Reset();
-		        var buildAndUpdateVisitor = new BuildAndUpdateDependenciesVisitor(dependeciesToBuild, projectsToUpdate);
-		        _algorithm.TraverseDependencies(buildAndUpdateVisitor, _options.Directory);
+            if (!_options.Dry)
+            {
+                return buildAndUpdateDependencies(dependeciesToBuild, projectsToUpdate);
+            }
+            else
+            {
+                printDryOutput(verifyVisitor.Changes, dependeciesToBuild, projectsToUpdate);
+                return checkArtifactsVisitor.ReturnCode;
+            }
+        }
 
-		        if (buildAndUpdateVisitor.ReturnCode == ReturnCode.Success)
-		        {
-			        if (buildAndUpdateVisitor.UpdatedPackages.Any())
-			        {
-				        _console.WriteLine(strings.UPDATED_PACKAGES);
-				        foreach (var package in buildAndUpdateVisitor.UpdatedPackages)
-				        {
-					        _console.WriteLine($"    {package}");
-				        }
-				        _console.WriteLine(strings.UPDATE_COMPLETE);
-			        }
-			        else
-			        {
-				        _console.WriteLine(strings.NOTHING_UPDATED);
-			        }
-		        }
+        private ReturnCode buildAndUpdateDependencies(List<string> dependeciesToBuild, List<string> projectsToUpdate)
+        {
+            _algorithm.Reset();
+            var buildAndUpdateVisitor = new BuildAndUpdateDependenciesVisitor(dependeciesToBuild, projectsToUpdate);
+            _algorithm.TraverseDependencies(buildAndUpdateVisitor, _options.Directory);
 
-		        return buildAndUpdateVisitor.ReturnCode;
-	        }
-	        else
-	        {
-		        VerifyCorrectBranchDryVisitor dryVisitor = verifyVisitor as VerifyCorrectBranchDryVisitor;
-		        if (dryVisitor == null)
-		        {
-			        return ReturnCode.InvalidArguments;
-		        }
+            if (buildAndUpdateVisitor.ReturnCode == ReturnCode.Success)
+            {
+                if (buildAndUpdateVisitor.UpdatedPackages.Any())
+                {
+                    _console.WriteLine(strings.UPDATED_PACKAGES);
+                    foreach (var package in buildAndUpdateVisitor.UpdatedPackages)
+                    {
+                        _console.WriteLine($"    {package}");
+                    }
+                    _console.WriteLine(strings.UPDATE_COMPLETE);
+                }
+                else
+                {
+                    _console.WriteLine(strings.NOTHING_UPDATED);
+                }
+            }
 
-				_console.WriteLine(strings.BRANCH_CHANGES);
-				if (dryVisitor.Changes.Count > 0)
-				{
-					foreach (string entry in dryVisitor.Changes)
-					{
-						_console.WriteLine($"\t{entry}");
-					}
-				}
-				else
-				{
-					_console.WriteLine($"\t{strings.DEPS_CORRECT_BRANCH}");
-				}
+            return buildAndUpdateVisitor.ReturnCode;
+        }
 
-				_console.WriteLine(strings.DEPENDENCIES_TO_BUILD);
-				if (dependeciesToBuild.Count > 0)
-				{
-					foreach (string entry in dependeciesToBuild)
-					{
-						_console.WriteLine($"\t{entry}");
-					}
-				}
-				else
-				{
-					_console.WriteLine($"\t{strings.PACKAGES_UP_TO_DATE}");
-				}
+         private void printDryOutput(List<string> branchesToUpdate, List<string> dependeciesToBuild, List<string> projectsToUpdate)
+        {
+            _console.WriteLine(strings.BRANCH_CHANGES);
+            if (branchesToUpdate.Count > 0)
+            {
+                foreach (string entry in branchesToUpdate)
+                {
+                    _console.WriteLine($"\t{entry}");
+                }
+            }
+            else
+            {
+                _console.WriteLine($"\t{strings.DEPS_CORRECT_BRANCH}");
+            }
 
-				_console.WriteLine(strings.PROJECTS_TO_UPDATE);
-				if (projectsToUpdate.Count > 0)
-				{
-					foreach (string entry in projectsToUpdate)
-					{
-						_console.WriteLine($"\t{entry}");
-					}
-				}
-				else
-				{
-					_console.WriteLine($"\t{strings.PROJECTS_UP_TO_DATE}");
-				}
+            _console.WriteLine(strings.DEPENDENCIES_TO_BUILD);
+            if (dependeciesToBuild.Count > 0)
+            {
+                foreach (string entry in dependeciesToBuild)
+                {
+                    _console.WriteLine($"\t{entry}");
+                }
+            }
+            else
+            {
+                _console.WriteLine($"\t{strings.PACKAGES_UP_TO_DATE}");
+            }
 
-		        return dryVisitor.ReturnCode;
-	        }
+            _console.WriteLine(strings.PROJECTS_TO_UPDATE);
+            if (projectsToUpdate.Count > 0)
+            {
+                foreach (string entry in projectsToUpdate)
+                {
+                    _console.WriteLine($"\t{entry}");
+                }
+            }
+            else
+            {
+                _console.WriteLine($"\t{strings.PROJECTS_UP_TO_DATE}");
+            }
         }
 
         #endregion
